requirements.txt
Pillow==9.5.0
Requests==2.30.0
streamlit==1.22.0
transformers==4.29.2
